feat: create Handlers Module - move all handlers to new module
- Create ClubDoorman/Services/Handlers/ folder
- Move UpdateDispatcher, IntroFlowService, BotPermissionsService, MessageHandler, CallbackQueryHandler, ChatMemberHandler
- Update namespace to ClubDoorman.Services.Handlers
- Create HandlersModule.cs with AddHandlersServices() extension method
- Remove all direct handler registrations from Program.cs (6 huge constructions)
- Program.cs reduced from 420 to 337 lines (-83 lines, -20%)
- All tests pass successfully
- Update checklist progress to 100% (4/4 phases completed)
